requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
flask
pypdf
python-dotenv
pyttsx3
gtts
playsound

